HW6 - fix notices 1
